Cache resolution masks per detector/panel
Because of implementation, generating a panel mask based on resolution
is a relatively slow process, as it works by naively calculating the resolution
for every pixel on a per-pixel basis. In normal spotfinding this object is
reused with a local cache and so this isn't run into. With per-image-analysis
spotfinding with find_spot_server this isn't the case, so we hit this
expense for every call.

This patch adds a cache round the call to generate the masker, which means that
although this expense is still paid, it's only paid the first time a particular
(panel, beam) combination is used.

The cache is deliberately set small to avoid excessive memory usage; this will
cause problems if repeatedly switching between N detectors but should be fine
under the assumption that successive images through the per-image-analysis
are mostly the same (panel, beam).

Fixes #1170, although not the root cause.