Support imagesets of len > 1 (e.g. grid scans) in dials.stills_process (#1174)
`dials.stills_process` needs something atomic to split on to send to the ranks. In the case of a grid scan of stills, `ExperimentListFactory` returns 1 experiment with an `ImageSequence` of length N. This changes that behavior so that in the `do_import` function of `dials.stills_process`, the single experiment is converted to a list of experiments of length N, each with an imageset of length 1. Then `dials.stills_process` can split these off to the ranks.

Still fast because load_models=False on import.