Change default for masking parameter 'untrusted.panel' from 0 to None
If panel is not specified, use panel 0.

Reason: due to how phil works with multiple=True, if the default for untrusted.panel is 0, then the following phil file results in an empty list of untrusted regions:
untrusted {
  panel = 0
}
If instead the default for untrusted.panel is None, the above phil works as expected, with the list of untrusted regions containing one item, with panel==0.