Change memory limit detection logic
The memory check logic appears to have been the following:
* We calculate the amount of memory required by one process
* We take the available system memory and multiply it by a factor
  of 0.75 (parameter 'integration.block.max_memory_usage')
* We divide the first second number by the first number, and use
  the rounded down value of that division to limit the number of
  processes.

This is probably fine, but it builds on the assumption that everything
else running on your computer fits inside 'system memory * (1 -
integration.block.max_memory_usage) + swap size'.
This is because "available system memory" literally means "theoretically
available physical memory in this machine". It does not mean "memory
available to allocate". Incidentally it also excludes swap memory.

I think we need to do a bit better with the memory estimation.
This commit uses the following logic:
* Find the amount of memory available for allocation
  (psutil.virtual_memory().available) and the amount of memory available
  in swap (psutil.swap_memory().free)
* Subtract the second from the first to obtain the amount of
  memory available for immediate allocation, multiply it by
  'integration.block.max_memory_usage'
* Divide this by the amount of memory required by one process,
  round down and limit the number of processes to this number.
* If this number falls below 1 then
  * Check whether the amount of memory available for immediate
    allocation including swap is greater than the amount of memory required
    by one process
  * If this is the case, then continue with 1 process
  * Otherwise crash out with a MemoryError (rather than RuntimeError)

This also writes out all the obtained numbers into the log.