Add best_monoclinic_beta=True to dials.refine_bravais_settings and dials.symmetry (#1226)
This is similar to the [POINTLESS `SETTING` parameter](http://www.ccp4.ac.uk/html/pointless.html#setting) and allows the user to enforce that C2 is selected in preference to I2. The default behaviour remains as before, i.e. I2 is preferred to C2 if it gives rise to a less oblique cell.

Fixes #1222, fixes #1224