Fix blank "Unable to handle" messages in OptionParser
We changed the error handling for unprocessed arguments at some point.
This means that previously when unhandled arguments were processed we
had some context for why they failed. Since we don't have that at the
moment, print the argument on it's own - this is more information than
nothing.

Fixes #600. Fixes #522. Fixes #563.