Add facility to ignore Python3-only tests, #1175
If a test fails on Python 2 and is expected to do so then mark it with
the 'python3' fixture.

If a test fails during the pytest discovery/collection phase due to a
SyntaxError in the test file, then add the filename to the exclusion
list in conftest.py. If the SyntaxError happens due to an import in the
test file then consider moving the import into the test function first.