Improvements to image stacking (#1302)
Add new features to image stacking:
* max
* mean
* sum
Modes, with the former two maintaining a sensible scale on the images. Also add a selector pull-down to select preferred stack type. Set max mode as default since that could well be more useful to the average user

Disable the "Kabsch" buttons if any of these are selected, since the dispersion analysis etc. does not make sense if the images have been manipulated. Also reset the settings.display parameter to "image" in case we were in another view when entering stack_images mode.

Also adjust max brightness -> 1000 to help with looking at summed images. 

As a side-effect always treat the data as doubles, to ensure the tools work correctly for raw and corrected data. 

Co-authored-by: Richard Gildea <richard.gildea@diamond.ac.uk>