Set d_min=None if interpolation error
If the fitted curve doesn't intersect with the desired limit, then
the previous behaviour defaulted to setting d_min as the high
resolution limit extent of the input data. In some cases, this
would lead to all data being used in subsequent analysis. Better
instead to be explicit that we have failed to determine a resolution
estimate and set d_min=None.