Resolution analysis: set d_min=None if interpolation error (#1378)
* Set d_min=None if interpolation error

If the fitted curve doesn't intersect with the desired limit, then
the previous behaviour defaulted to setting d_min as the high
resolution limit extent of the input data. In some cases, this
would lead to all data being used in subsequent analysis. Better
instead to be explicit that we have failed to determine a resolution
estimate and set d_min=None.

* Don't override default nbins for resolution analysis

20 bins is insufficient if the true resolution limit of the data
is much lower than the resolution limit to which data were integrated.
Better to stick with the default nbins=100.

* Scaling fix for CC½ fit returning None