Change behavior of reference_geometry to sync the detector instead of overwritting it
Overwriting the detector means changes in the format class that don't affect the panel positions are lost.  Case in point, the original reference geometry for the SACLA MPCCD test that is part of test_stills_process has a gain of 1 and a high saturation value.  The current format class indicates the gain is 10 and caps the trusted range lower.  To use these values and still use the refined panel positions, it is better to sync recursively the detector frames than to overwrite the detector.

This causes the spotfinding parameters for the SACLA MPCCD test to change and actually integrates all 4 images (instead of 3) from the test set.