Change default separate_panels=False
When separate_panels=True, outlier rejection is performed for each panel
separately prior to refinement. This is appropriate when we expect the
data on each panel to be drawn from different distributions (for example
prior to metrology corrections). However, it causes a large increase in
the execution time of refinement - dominating the overall job time for
a test case using data from I23.

This parameter should be set to true when doing metrology refinement,
but for normal data processing use, false is likely to be acceptable.