Take into account phi in nearest neighbour analysis
In the "local" index assignment method, the nearest neighbour
analysis should only identify as neighbouring those reflections
that are close in phi. If, for example, there is a rotation misset
in a scan that is more than one full rotation, this previously
could cause issues for the local index assignment, leading to low
numbers of indexed reflections.

Fixes #1458