Default to choosing the cluster that contains the most identity reindexing ops
This can be important if the lattice symmetry is only very approximately related by
pseudosymmetry to the true symmetry. If all potential reindexing ops are genuine
indexing ambiguities, then it doesn't matter which one is chosen, however if not,
then choosing the wrong one will distort the true unit cell. In such cases it is
likely that the input datasets were already indexed consistently, therefore default
to choosing the cluster that contains the most identity reindexing ops.