Reword newsfragment for a lay reader
Remember the audience!  In the PHIL scope, indexing strategies are
referred to as 'indexing methods', so use that terminology here.  Also,
the words 'PHIL scope' are meaningless to most users, use
'configuration options' instead.  Avoid the imperative mood in case it
is mistaken as an instruction to the user to change the way they use
dials.index — after all, a newsfragment is not a commit message.