dials.merge: Fix output of SigF, N+, N- in merged.mtz
For anomalous=True, when merging anomalous Fs to normal Fs,
dont use_internval_variances. This gives an incorrect
(overly high) value for SigF.

Also, the miller index multiplicities were being written as N+,
N- however this was intended to be the redundancy of observations
of each symmetry group, which is fixed here.