Address comments from @ndevenish
was completely correct this had introduced errors, had things back to front,
so fixed the doc string to what it should be and added extra test which
exercises this as well as fixing the previous test. Now I think this is
correct. Also removed some of the shell game nonsense by using sensible
internal variable names (xyz, ref)