Hack to avoid call to nullify_format_instance() (#1539)
The call to nullify_format_instance() is only
necessary for HDF5 when we are using multiprocessing.
Here, we are using a single processor, so it only adds
unnecessary overhead. Avoiding the call can save ~0.15
seconds per image when this function is called via the
dlstbx PIA service (~10% of the runtime)

See also https://github.com/dials/dials/issues/144