Fix image viewer problems with multiple stills (#1463)
Fixes various bugs that manifest in the image viewer when viewing stills. In particular, combined experiments can now be parsed with reflections placed on the correct image. Further, when paging through multiple images, I noticed that the caching of the dispersion_debug_memo was seemingly broken for my use case, so I simply bypassed it for viewing stills. I think with these commits, the image viewer is functional the way I expect it should be for stills.