Remove unnecessary get_raw_data() call
The mask generation code doesn't actually use the image data, but
just uses the image to get the panel sizes - which are available in
the panel object! This extra call to imageset.get_raw_data() for the
first image of an imageset causes significant overhead e.g. to the
per-image analysis service used at DLS.

Fixes 1449