Fix test failure on Linux by using different random seed
This appears to be a consequence of using np.random.rand in place
of flex.random_double to initialise the starting coordinates. As
a result the automatic determination of number of dimensions chooses
7 instead of 6 dimensions, which seems to tip the balance from
success to failure. This highlights that this probably isn't the
most robust test case - the resulting likelihood for the correct
space group isn't particularly high, even when the test is working.