Also cache undefined Rij values
If we know that the correlaton coefficient is undefined for this particular
combination of data sets then we can cache that knowledge and avoid the cost
of rediscovering this fact for other symmetry-related combinations of the
same data sets.

This shouldn't change any results, and will only likely have any significant
performance benefit when dealing with many sparse data sets (i.e. where there
are many pairs of data sets where it isn't possible to calculate a valid
correlation coefficient).

Co-authored-by: Daniel Paley <dwpaley@lbl.gov>