Drop 'set' command, as it breaks the pipeline
For whatever reason this leads to the stage outcome to not be set to
'success' although the job outcome is shown as successful either way.
This is probably an Azure bug, but given my experience with more
complicated issues (ie. the broken cache issue) it's not worth reporting
as it won't get beyond first level support.