Add '--prebuilt-cctbx' option to bootstrap script (#1706)
With this option set bootstrap no longer attempts to build
all of cctbx from scratch. As using a prebuilt cctbx package
complicates a development installation this is not currently
recommended outside of one-shot builds such as on Azure.

Further, this commit only prepares the bootstrap script, and
doesn't enable its use on Azure yet.