Run xfel-regression tests on Azure (#1707)
This restructures the Azure pipelines into three stages. Syntax validation
and flake8 checking are merged together into a single step in the first stage.
If that succeeds we build and test dials on various platforms in the second
stage. On pull request builds only this now also includes xfel-regression tests.
The results of the xfel-regression tests are not reported at this point though,
they are collated in a separate job in the final stage.

Closes #1671