Restore cache, but more cleanly
By simply computing a hash of the settings including the imageset this still
fixes the bug whilst leaving the other thing still fixed. Changing a request
dict to a simple hash of a tuple of the settings stops a future reader
from working through the code looking for where that request is sent to.

Also named the hash the same as the object which is hashed, to make the
connection clearer.

Bug still fixed, no previous bugs restored. I believe these are good changes.