Fix image_viewer display of spotfinding intermediates for stills (#1761)
By simply computing a hash of the settings including the imageset this
still fixes #1734 whilst leaving #1213 fixed. Changing a request dict
to a simple hash of a tuple of the settings stops a future reader from
working through the code looking for where that request is sent to.

Also named the hash the same as the object which is hashed, to make the
connection clearer.