Change text contrast colours in image_viewer (#1781)
The resolution rings and Miller index text overlays switched between
white and black to provide contrast with the image, but the basis vector
labels were always black. Fix that by providing a function to choose a
suitable text colour and use that in each place.

Also, change to light vs dark grey rather than white vs black. The
extremes (white and black) are common for pixels in diffraction images
where the contrast is tuned to a narrow part of the full dynamic range.
Better to choose a colour that is still visible even against a solid
white or black background.