Changes to try to improve spot finding for ED
1. Make the erosion less severe before calculating mean of background
2. Use the original dispersion mask for calculating strong pixel mask

The algorithm works by:
1. computing the local mean and variance at each pixel
2. computing the index of dispersion at each pixel
3. calculating a dispersion mask where the variance  / mean is > 1.
4. the dispersion mask shows potential strong pixels in the neighbourhood
5. the mask is then eroded slightly
6. the mean is then calculated ignoring the potential strong pixels under eroded dispersion mask
7. the strong spot mask is then calculated by looking at pixels greater than the mean (which is now an esimate of the local background).

The rationale for these changes are as follows:

1. During the background mean calculation step (6). We ideally only want
to consider pixels which are background. If we erode too far then we end
up using all the pixels which defeats the purpose of the second mean
calculation step.

2. During the final threshold step (7) if we use the eroded mask we have
fewer potential strong pixels to choose from. We are essentially
selecting only the peak if anything. It is probably better to use the
original uneroded dispersion mask and select strong pixels from within
that in the final step.