Add SparseFlex.select and test
This turned out to be quite complicated. The intersection and
intersection_i_seqs methods did not quite do the right thing
because they require sorted input arrays, whereas this has to
preserve the sorted order. I ended up with a loop in Python
to do what I want, but this should be converted to C++ for
speed.