Update polarisation normal vector when refining a Beam
Typically, the beam direction is refined only in the X-Z plane,
while the polarisation normal vector is along Y, so in most cases
for MX this makes no differnce. However the beam can be refined
around two angles, and in those cases would have become
inconsistent before this change. Fixes #1939