Experimental dev.dials.ssx_index: Index a block of ssx data (#1955)
This script wraps calls to the dials.index code, to attempt indexing
with the fft1d and real_space_grid_search methods if appropriate.

Processing is distributed across processes using concurrent.futures and
then the results are gathered and summary statistics generated, which
are plotted in an output html report. Unit cell clustering is also run
and reported, and the beam and detector models are combined to allow
continued processing with dials.refine.