Revert to not using np.intersect1d
The NumPy method returns sorted values, which means that it only
works for selection when the input arrays are themselves sorted.
This is often the case but cannot be guaranteed. Therefore this
method is not suitable for use in SparseFlex.select. Added a
test to demonstrate the failure of this method when the input
arrays are unsorted.