Add new spotfinding algorithm, threshold=radial_profile  (#2009)
This method calculates background value and sigma in 2θ shells, then sets
a threshold at a level n_sigma above the radially-averaged background.
As such, it is important to have the beam centre correct and to mask out
any significant shadows. The method may be particularly useful for
electron diffraction images, where there can be considerable inelastic
scatter around low resolution spots, and it is difficult to tune
parameters of the dispersion algorithms (see e.g. #1803). The algorithm
is relatively insensitive to noise properties of the detector, which
helps for the case of integrating detectors with poorly known gain and
response statistics.

The description of 'peakfinder 8' in https://doi.org/10.1107/S1600576714007626
was helpful in the development of this method, however the code is based
on the approach taken in dials.background.

A future implementation will have the imageset, panel number and ROI
definition, all of which are needed to identify exactly which pixels we
are trying to threshold.

Co-authored-by: Richard Gildea <rjgildea@users.noreply.github.com>