find_spots/import: No __main__ check, deprecate ImageImporter (#2080)
Both dials.find_spots and dials.import check if they are imported as
__main__ to decide whether to configure logging (and thus show any
output). This assumption fails if running from a standard
console_scripts entry point.

This refactors both scripts away from the legacy 'a = Script(); a.run()'.

Because Import is implicitly done when parsing arguments, and it does a
first pass to determine the logging level, the function is quite
intertangled with the ArgumentParser flow. Thus, I've added a
configure_logging= function parameter which toggles this.

The ImageImporter class is kept, but deprecated. I've updated the tests
that used this directly.

There are also tests that assume that the find_spots.run function
returns the experiments and reflection table, and passes in
argument-style parameters. For this reason, find_spots.run here returns
if passed return_results=True. The tests should probably be refactored
into running as-a-subprocess for this.