Always insert dials into the path when refreshing (#2089)
In cases where prebuilt cctbx was being used, the modules/ folder might not be
a tbx-repository, and so the "import dials" inside dxtbx would fail to even
find DIALS as a namespace repository. This meant that there was no broken
namespace module in sys.modules.

With this change, the dials/src folder is always added to sys.path while
refreshing.