Protect against mean=0 in simulation poisson distribution (#2094)
Fix against case in testing where a poisson distribution with zero mean
could be generated. This works for bootstrap builds and conda-forge
prebuilt cctbx on linux, but prebuilt on mac fails.

Presumably this is sensitive to how scitbx was compiled - if boost
assertions are enabled when scitbx is compiled, it can encounter this
assert, killing the process.