image_viewer: Cast to int in pyslip
In conda-forge releases, this was causing an uncaptured exception that
resulted in a SEGV from wxPython. I don't know why this doesn't happen
on development releases, but the function is documented as only accepting
an int anyway.

This does not fix the other longstanding issue of SEGV on exit on MacOS.