Remove failing rs_mapper test
This has been failing since cctbx/dxtbx#455 was merged, however I
don't think the test makes sense as is, either before or after these
changes. dials/dxtbx does not handle screening data collections from
eigers gracefully - before this change it interprets as 3 x 45-degree
images (so the starting angle is correct, but the oscillation range
of each image is wrong), and after as 3 x 0.15-degree oscillations
(so the starting angles are incorrect, but the oscillation ranges
are correct). See also cctbx/dxtbx#13.