Do not crash when there are fewer than two parameters to constrain
Previously, constraint scopes created for parameterisations that
were removed due to having too few reflections would cause a
crash. Now this elicits a warning and a None constraint, which
is later removed from the manager.

In addition, the ability to constrain scan-varying parameterisations
was removed as this was untested and apparently buggy.