Constraints mapping bug (#2155)
* Fix bug in matching models by experiment id in constraints.

The index of the experiment id was being used, not the index of
the model. So, if id=5000 is an experiment using the second
detector, the previous behaviour tried to apply constraints to
the 5001st detector 5001, not the 2nd detector.

* Do not crash when there are fewer than two parameters to constrain

Previously, constraint scopes created for parameterisations that
were removed due to having too few reflections would cause a
crash. Now this elicits a warning and a None constraint, which
is later removed from the manager.

In addition, the ability to constrain scan-varying parameterisations
was removed as this was untested and apparently buggy.