Reindex refls_for_sym to reference setting
Currently in dials.symmetry, if the Laue group is
set to Auto, and the supplied data are not in the
canonical setting for a given space group, they
are not reindexed before being passed to
`run_systematic_absences_checks`. This leads to
a failure in identifying the appropriate screw
axes. This commit adds an additional call to
`_reindex_experiments_reflections` which should
fix this (bug).