Add known orientations to dials.stills_process (#2110)
* Add known_orientations to dials.stills_process

An old feature of xtc_process, if the user has previously indexed the images, they can provide those crystal orientations using the known_orientations as a list of paths to experiment lists.  The bookkeeping is done by matching up filenames + image indices.

* Add require_known_orientation for known_orientations

If known_orientations are provided, and an orientation for an image is not found, this is whether or not to attempt to index the image from scratch using indexing.method

* Only try indexing with known_crystal_models if they are provided

* Handle multiple lattices in known_orientations for duplicate reflection filtering

This duplicate filtering code came when prepping Brewster 2018 Figure 5, where split spots on a detector 2.5m from the source were both indexed by known_orientations using the same miller_index.  That code didn't check for multiple lattices, hence this re-write.

* Check for no files in known_orientations paths

* Test known_orientations using indexing results from dials.stills_process

* Apply suggestions from code review

Removes early return for known_crystal_models so the later code that filters duplicate reflections can still run

Co-authored-by: Daniel Paley dwpaley@lbl.gov