Parallelise dials.rs_mapper (#2238)
Parallelise the calculation with concurrent.futures.

I checked that the output map is identical between this branch and
main for the dials.data insulin dataset. On the branch this was
split over 5 blocks and took about 3 seconds compared with 5
seconds on main. Memory requirements on the branch were a bit less
than double that on main (810 MB compared with 450 MB).

For the thaumatin dataset, the branch took 7s compared to 24s on
main, with the job split over 16 processes. Memory required went
up to 1.7 GB from 520 MB.