stills_process: reflection subsampling (#2275)
Add new feature to dials.stills_process: reflection_subsampling. If reflection_subsampling.enable=True, and an image fails to index, the program will randomly subsample the reflections and try again. It keeps trying, randomly subsampling in smaller amounts until a threshold is reached or the image successfully indexes.

The default is descending from 100% to 50% of the strong spots by 2% steps, randomly subsampling from the original pool of strong reflections at each step.

We have observed nearly doubling the number of indexed images for some samples using this approach, and we verified that the additional lattices are high quality (ie we are not adding merely weak lattices that failed to index).

* Add test for reflection_subsampling

The gain of the SACLA MPCCD detector is normally 10.  The new test wrong_gain changes it to 0.5 and produce more spots, which causes the 3rd image to fail to index. This is rescued by reflection_subsampling.

Co-authored-by: Daniel Paley dwpaley@lbl.gov
Co-authored-by: Asmit Bhowmick abhowmick@lbl.gov