Use C++14, unpin boost, and update CCTBX base (#2319)
Using C++14 by default exposes dials to failures on windows because of
the standard library overrides in cctbx, so remove those when building
dials libraries.

Using C++14 means that we can use Boost::Geometry in boost 1.78, which
dropped support for older versions. This means that we can build against
the January 2023 release of cctbx, which requires boost 1.78.

Additionally, dxtbx directly links to zlib but this was only ever pulled
into the conda environments implicitly, but on some configurations of
building dials on windows this is no longer the case - so make the
dependency explicit here.

Fixes #2291.