Add methods to Python: set_title, add_history and set_space_group_by_name.
In the case of set_space_group_by_name, inject into the boost.python class,
like how it is done for iotbx.mtz.object. But here we use a simpler method.
Is it enough? Depends on whether the space group symbol is uniquely
determined, and I suspect it is not...