dials.ssx_index: Improve results without unit cell values (#2372)
Currently, when indexing without a given unit cell, the max cell
estimation across all images is used. This tends to be very high
(~700A-1000A), and results in an overrepresentation of indexed cells
with very large cell values in the final reported cell clustering.

This PR changes this to use the 95th centile value from the distribution
of max unit cells, the use case when the cell values are not known.