Separate independent blocks of experiments in refinement (#2336)
Where the input experiment list contains groups of experiments that do not share models between the groups (i.e. disjoint sets), the new default behaviour is to refine each group separately. This reduces memory usage and run time. The results are not generally identical to the old behaviour, but the quality of the refined model is expected to be equivalently good.