Edit test_mtz_primitive_cell because now we again reindex back to the reference setting for MTZ export, so that files can be read by CCP4 programs.
See https://github.com/dials/dials/issues/1278 for some relevant
commentary. At that time, the change was made to export MTZ files
/without/ reindexing to reference, in order to fix inconsistencies.
It was noted that the files could not actually be read by Pointless,
however.

Now the reindex works correctly for I23, as the added test shows,
because MTZ files contain the same data whether exported from the
reference setting or reindexed to the primitive setting .