Fix RHEL8+ gltbx compilation to avoid system libc
On RHEL8 (and presumably other modern systems), gltbx compilation fails due to
a interaction between the conda-forge sysroot libc++ and the local system
libc++. This happens because gltbx adds the system include paths to the
build (multiple times).

This code detects this situation, and resolves it. DIALS bootstrap builds will
be built with the OpenGL CDT packages enabled.