Add mesa-libGL to linux bootstrap installations
This is to solve a longstanding issue where bootstrap builds would fail on
linux systems more recent than RHEL7.

Cause:

- The conda-forge sysroot stdlib system uses #include_next in order
  to chain to a wrapped stdlib.h.
- gltbx adds the system /usr/include and /usr/local/include multiple
  times to the include search path.
- When the c-f sysroot stdlib tries to #include_next, instead of the
  REAL sysroot stdlib.h, it picks up the system stdlib.h.
- This worked on RHEL7- the system stdlib was old enough that it
  apparently didn't use this mechanism.
- Since the RHEL8 system stdlib is recent enough to use this mechanism,
  it thinks it is being #included twice, and so hits the usual guards.

This "System include path" was added originally to work around the fact
that OpenGL is a complicated dependency and thus you need to use the
library from the system. This problem can also be solved using the
Conda-forge CDT system dependencies.