Fix ssx_index test
Before this stills_indexer outlier change, ssx_index had the default set to null, to match the stills_indexer requirement.  Now that that stills_indexer can be non-null, this test which uses the sequences indexer, needs to pass in null for the outlier rejeciton to match the original test data.