Allow bootstrap of release branches with CMake (#2727)
Run micromamba twice, if required. In cases where the requirements file
is explicit, AND we want to add extra dependencies, we need to do a
second pass, because conda will silently ignore the extra packages. We
want to add packages if we are running with prebuilt cctbx, and we have
explicit requirements on release branches.

Also, remove option for non-cmake prebuilt CCTBX. This was only ever
supported in CI, and we now use the CMake build for that purpose.

To do this, we have to allow release dependencies to be updated upon
install, for cmake. At the moment this isn't a directly supported
release combination, but we want to allow it for image building, so
let's accept the changes to that we are compatible with prebuilt CCTBX,
and take this into account next time.