Feature: add split= to dials.import (#2816)
Feature: add split= to dials.import

Syntax of split=start,end,size to split imported data into given chunk sizes. Also allows subset of the data to be used i.e. if start > 1 or end < last. If split=size used, without the start and end, it will just split the data from start to end in blocks of size images with the last one the remainder if the size does not divide evenly.

Fixes #2815

Much feedback from @biochem-fan in the process, thank you