Speed up tests (#2949)
* Speed up test_index.py::test_index_small_molecule_multi_sequence_3 by
setting a limit of the number of crystal models tried (10 rather than
50)

* This test was added in https://github.com/dials/dials/commit/55294a5be2e579869860dca7147f0a68bb0c8539
but there is no mention of what bug this refers to. The actual test appears to do very
little of any use, and is expensive because it runs everything from import to
export. I see no reason to keep this test.

* Set max_refine=5 to half the execution time of the test

* Why on Earth was this test set up to import and find spots on
a data set when all it needed was a stills .{expt,refl} pair?!