Consistent data preparation for `dials.symmetry` and `dials.cosym` (#3013)
These two programs used very similar code as part of the set up steps prior to symmetry analysis. This PR aligns the steps so that the programs can use the same function for set up, and then refactors to extract that function from the command line programs and puts it into a commonly imported module.