bootstrap: Check for shell, to not use bash on macos
If on macOS, then the shell is zsh with a very old bash available. An old
version of bash that the new mamba.sh uses features that are not present.
This is normally fine, because the user running zsh sources the dials script,
except this falls over at the bootstrap phase where indirect.sh is set to
use /bin/bash and is run directly (not as source).